+++ packages/expo-ui/src/community/pager-view/PagerView.tsx
@@ -7,7 +7,7 @@ import type { PagerViewProps } from './types';
  * paged view backed by Jetpack Compose's `HorizontalPager` on Android and
  * SwiftUI on iOS. Each child is treated as a separate page.
  */
-export function PagerView(_props: PagerViewProps): null {
+export function PagerView(_props: PagerViewProps): never {
   throw new Error(
     `@expo/ui/community/pager-view is not implemented on ${Platform.OS}. Supported platforms: android, ios.`
   );
