+++ packages/expo-ui/src/swift-ui/modifiers/scrollObservation.ts
@@ -1,7 +1,7 @@
 import { worklets } from '../../State/optionalWorklets';
 import { useWorkletProp } from '../../State/useWorkletProp';
 import { getStateId } from '../../State/utils';
-import type { ScrollGeometry } from '../ScrollView';
+import type { ScrollGeometry, ScrollPhase } from '../ScrollView';
 import { createModifier, createModifierWithEventListener } from './createModifier';
 
 /**
@@ -58,3 +58,27 @@ export function useScrollGeometryChange(callback?: (geometry: ScrollGeometry) =>
     callback(event)
   );
 }
+
+/**
+ * Fires when SwiftUI's scroll phase changes (e.g., the user begins dragging,
+ * the scroll view starts decelerating, or scrolling settles to idle). The
+ * second argument is the scroll geometry sampled at the phase transition,
+ * useful for reading the final offset on settle without subscribing to
+ * per-frame `onScrollGeometryChange`.
+ *
+ * Apply to a SwiftUI `ScrollView` (and other scrollable views). On iOS below
+ * 18.0 the modifier is a no-op.
+ *
+ * @platform ios 18.0+
+ * @platform tvos 18.0+
+ *
+ * @see Official [SwiftUI documentation](https://developer.apple.com/documentation/swiftui/view/onscrollphasechange(_:)).
+ */
+export const onScrollPhaseChange = (
+  callback: (phase: ScrollPhase, geometry: ScrollGeometry) => void
+) =>
+  createModifierWithEventListener(
+    'onScrollPhaseChange',
+    (event: { phase: ScrollPhase; geometry: ScrollGeometry }) =>
+      callback(event.phase, event.geometry)
+  );
